+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Features
 - Load self-hosted fonts to replace remote fonts
+- Add mermaid support
+- Add dark/light mode for code blocks 
 
 ## Bugs
 - Fix error in icons and borders